+++ frontend/src/components/Workspace/FlowChart/FlowChartNode/FileSelect.tsx
@@ -16,8 +16,8 @@ import { IconButton, Tooltip, Typography } from "@mui/material"
 import ButtonGroup from "@mui/material/ButtonGroup"
 
 import { FILE_TREE_TYPE, FILE_TREE_TYPE_SET } from "api/files/Files"
+import { CsvParamSettingDialog } from "components/Workspace/FlowChart/Dialog/CsvParamSettingDialog"
 import { DialogContext } from "components/Workspace/FlowChart/Dialog/DialogContext"
-import { CsvParamSettingDialog } from "components/Workspace/FlowChart/FlowChartNode/CsvParamSettingDialog"
 import { LinearProgressWithLabel } from "components/Workspace/FlowChart/FlowChartNode/LinerProgressWithLabel"
 import { FileNodeFactory } from "factories/FileNodeFactory"
 import { getFilesTree } from "store/slice/FilesTree/FilesTreeAction"
